+++ railties/lib/rails/generators/rails/app/app_generator.rb
@@ -366,6 +366,13 @@ module Rails
       end
       remove_task :update_bin_files
 
+      def update_action_text
+        unless skip_action_text?
+          rails_command "action_text:update", inline: true
+        end
+      end
+      remove_task :update_action_text
+
       def update_active_storage
         unless skip_active_storage?
           rails_command "active_storage:update", inline: true